[painter/gl] move #glDrawLine() to common code
It no longer depends on the GL implementation.
