+++ internal/painter/gl/painter.go
@@ -4,6 +4,7 @@ package gl
 import (
 	"fmt"
 	"image"
+	"image/color"
 	"image/draw"
 	"math"
 
@@ -124,6 +125,30 @@ func (p *painter) freeTexture(obj fyne.CanvasObject) {
 	p.ctx.DeleteTexture(Texture(texture))
 	p.logError()
 	cache.DeleteTexture(obj)
+}
+
+func (p *painter) glDrawLine(width float32, col color.Color, feather float32) {
+	p.ctx.UseProgram(p.lineProgram)
+
+	p.ctx.BlendFunc(srcAlpha, oneMinusSrcAlpha)
+	p.logError()
+
+	colorUniform := p.ctx.GetUniformLocation(p.lineProgram, "color")
+	r, g, b, a := col.RGBA()
+	if a == 0 {
+		p.ctx.Uniform4f(colorUniform, 0, 0, 0, 0)
+	} else {
+		alpha := float32(a)
+		p.ctx.Uniform4f(colorUniform, float32(r)/alpha, float32(g)/alpha, float32(b)/alpha, alpha/0xffff)
+	}
+	lineWidthUniform := p.ctx.GetUniformLocation(p.lineProgram, "lineWidth")
+	p.ctx.Uniform1f(lineWidthUniform, width)
+
+	featherUniform := p.ctx.GetUniformLocation(p.lineProgram, "feather")
+	p.ctx.Uniform1f(featherUniform, feather)
+
+	p.ctx.DrawArrays(triangles, 0, 6)
+	p.logError()
 }
 
 func (p *painter) imgToTexture(img image.Image, textureFilter canvas.ImageScale) Texture {
